Rename parser.js to index.js
